+++ testData/org/elixir_lang/inspection/no_parentheses_many_strict_test_case/SingleNestedKeywordValue.ex
@@ -1,8 +1,4 @@
-# A noParenthesesExpression can't itself have a noParenthesesExpression as a keyword value according to
-# elixir_parser.yrl, but it is allowed so that {@link: org.elixir_lang.inspection.NoParenthesesManyStrict} can mark it
-# as an error
+# A no-parentheses call with several arguments is a valid keyword value of another no-parentheses call.
 first_function first_positional,
-               # `second_function second_position, ambiguous_keyword_key: ambigious_keyword_value` is the errant
-               # nested noParenthesesExpression, but the actual error will appear on the first `,`.
-               first_keyword_key: second_function second_positional<error descr="unexpected comma.  Parentheses are required to solve ambiguity in nested calls.">,</error>
+               first_keyword_key: second_function second_positional,
                   ambiguous_keyword_key: ambiguous_keyword_value